Move the vfs/xattr functions from handle to node.
The fuse layer never calls the xattr functions with a handle, meaning
that even if the file is open we don't know that. So all operations that
require an open file must open the file explicitly.

Additionally, the only currently supported attribute (the hash of the
file) does not require an open file handle to retrieve.

This simplifies the implementation by removing a layer of indirection.